Finalize switchover to new implementation
