+++ core/src/main/scala/kafka/coordinator/transaction/TransactionMarkerChannelManager.scala
@@ -19,7 +19,7 @@ package kafka.coordinator.transaction
 
 import java.util
 import java.util.concurrent.{BlockingQueue, ConcurrentHashMap, LinkedBlockingQueue}
-import kafka.common.{InterBrokerRequestManager, GenericInterBrokerSendThread, RequestAndCompletionHandler}
+import kafka.common.{InterBrokerRequestManager, InterBrokerSendThread, RequestAndCompletionHandler}
 import kafka.server.{KafkaConfig, MetadataCache, RequestLocal}
 import kafka.utils.Implicits._
 import kafka.utils.{CoreUtils, Logging}
@@ -65,7 +65,7 @@ class TxnMarkerQueue(@volatile var destination: Node) {
 class TransactionMarkerChannelManager(
                                        config: KafkaConfig,
                                        metadataCache: MetadataCache,
-                                       interBrokerSendThread: GenericInterBrokerSendThread,
+                                       interBrokerSendThread: InterBrokerSendThread,
                                        txnStateManager: TransactionStateManager,
                                        time: Time
 ) extends InterBrokerRequestManager(interBrokerSendThread, 1)
